Basic Espionage Screen. Dynamic hometab screen.
There is a basic espionage screen with the currently discussed panels on
it. There is no backend work does so far. Just the View.

The BobisbackHomeTabPanel is a panel that is created on the bottom of the
home tab that will only display when the hometab is active. This panel
will house all buttons that are added by my mod. The buttons can be made
by calling the addButton function. This allows dynamic creation of buttons
without changing any code. There is no base game code modified with this
class. The whole point is to have no base game code changed. This is to
create compaibility between other mods and this one.